+++ client/changes.md
@@ -1,5 +1,7 @@
 Changes
 
+1.0.0 (45)
+
 - public bug products
 - pdf in new tab
 - delete products
@@ -13,3 +15,12 @@ Changes
 - porto message ?
 - seichern -> speichern
 - divider for news
+
+1.0.0 (46)
+
+- Content for the rest of the pages...
+- PDF uploader
+- Porto bug auf Produkt-Detailseite
+- Bug on products without image
+- Placeholder image under product details
+- Porto bug on cart without any product with porto